requirements.txt
pandas
numpy
scikit-learn
sentence-transformers
google-generativeai
python-dotenv
tqdm